chore(readme): Add readme about Linux, Chromium, and NotAllowedError workarounds (#4)
The k810 fn config tool/WebHID api appears to raise `NotAllowedError` errors on Linux (e.g. ubuntu) and also with the ubuntu Snap Chromium browser. A section is added to Readme to outline the steps with references that resolve/work around the error on Linux systems and ubuntu Snap Chromium browser.